requirements.txt
fastapi
uvicorn
python-dotenv
numpy
pydantic
scikit-learn

